chore: move new upgrade behind --improved flag
